pyproject.toml
[project]
name = "e-commerce-personal-shopper-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deep-translator>=1.11.4",
    "fastapi==0.117.1",
    "ipykernel>=6.0.0",
    "langchain==0.3.27",
    "langchain-community==0.3.29",
    "langchain-google-genai==2.1.12",
    "langgraph==0.6.7",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "pydantic==2.11.9",
    "pydantic-settings==2.10.1",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "qdrant-client==1.15.1",
    "selenium==4.35.0",
    "sentence-transformers>=2.2.0",
    "sqlalchemy==2.0.43",
    "sse-starlette==3.0.2",
    "torch>=2.0.0",
    "webdriver-manager>=4.0.2",
]
